# add AUTOGOLD_PROFILE for profiling performance