e.g. for cases like https://github.com/sourcegraph/sourcegraph/pull/18189 this helps to debug where time is being spent:

```
$ AUTOGOLD_PROFILE=t go test -v -run='TestUnionMerge/#02' ./cmd/frontend/graphqlbackend/
=== RUN   TestUnionMerge
=== RUN   TestUnionMerge/#02
autogold: profile:
  getPackageNameAndPath: 289.5897ms
  stringify (want):      1.267376868s
  stringify (got):       1.377507419s
  diffing   (got):       56.162µs
  acquire path lock:     0s
  rewrite autogold.Want: 0s
--- PASS: TestUnionMerge (2.93s)
    --- PASS: TestUnionMerge/#02 (2.93s)
PASS
ok  	github.com/sourcegraph/sourcegraph/cmd/frontend/graphqlbackend	3.380s
```

Signed-off-by: Stephen Gutekanst <stephen@hexops.com>